+++ core/src/db/insert/MemManagerImpl.cpp
@@ -62,10 +62,12 @@ MemManagerImpl::InsertVectors(const std::string& collection_id, int64_t length, 
                               const uint8_t* vectors, uint64_t lsn, std::set<std::string>& flushed_tables) {
     flushed_tables.clear();
     if (GetCurrentMem() > options_.insert_buffer_size_) {
-        ENGINE_LOG_DEBUG << "Insert buffer size exceeds limit. Performing force flush";
+        ENGINE_LOG_DEBUG << LogOut("[%s][%ld] ", "insert", 0)
+                         << "Insert buffer size exceeds limit. Performing force flush";
         // TODO(zhiru): Don't apply delete here in order to avoid possible concurrency issues with Merge
         auto status = Flush(flushed_tables, false);
         if (!status.ok()) {
+            ENGINE_LOG_DEBUG << LogOut("[%s][%ld] ", "insert", 0) << "Flush fail: " << status.message();
             return status;
         }
     }