+++ core/src/utils/Log.cpp
@@ -1,0 +1,30 @@
+// Copyright (C) 2019-2020 Zilliz. All rights reserved.
+//
+// Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance
+// with the License. You may obtain a copy of the License at
+//
+// http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software distributed under the License
+// is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
+// or implied. See the License for the specific language governing permissions and limitations under the License.
+#include "utils/Log.h"
+
+#include <cstdarg>
+#include <cstdio>
+#include <memory>
+#include <string>
+
+namespace milvus {
+
+std::string
+LogOut(const std::string& pattern, ...) {
+    auto str_p = std::make_unique<char[]>(pattern.length() + 256);
+    va_list vl;
+    va_start(vl, pattern.c_str());
+    vsprintf(str_p.get(), pattern.c_str(), vl);
+    va_end(vl);
+    return std::string(str_p.get());
+}
+
+}  // namespace milvus