+++ CHANGELOG.md
@@ -24,6 +24,7 @@ Please mark all change in change log and use the issue from GitHub
 -   \#1784 Add Substructure and Superstructure in http module
 -   \#1858 Disable S3 build
 -   \#1882 Add index annoy into http module
+-   \#1886 Refactor log on search and insert request
 
 ## Task